minor refactoring

+++ include/textbox.hpp
@@ -12,17 +12,17 @@ class Textbox
 {
 public:
     Textbox();
-    Textbox(int l_visible, int l_charSize, int l_width, sf::Vector2f l_screenPos);
+    Textbox(size_t l_visible, int l_charSize, int l_width, sf::Vector2f l_screenPos);
     ~Textbox();
 
-    void Setup(int l_visible, int l_charSize, int l_width, sf::Vector2f l_screenPos);
+    void Setup(size_t l_visible, int l_charSize, int l_width, sf::Vector2f l_screenPos);
     void Add(std::string l_message);
     void Clear();
 
     void Render(sf::RenderWindow& l_window);
 private:
     MessageContainer m_messages;
-    int m_numVisible;
+    size_t m_numVisible;
 
     sf::RectangleShape m_backdrop;
     sf::Font m_font;
